requirements.txt
mysql-connector == 2.2.9
pillow
flask == 2.2.5
flask_restx == 1.1.0
pydantic == 2.1.1
flask_sqlalchemy == 3.0.5
Flask_Cors == 4.0.0
matplotlib == 3.7.2
boto3 == 1.28.23
requests~=2.31.0